remove not-very-useful MetronomeControls component
